refactor: shared validation factory
