+++ packages/typescript-fetch-runtime/src/zod.ts
@@ -1,50 +1,15 @@
 import type {z} from "zod"
-import type {Res, StatusCode} from "./main"
+import {responseValidationFactoryFactory} from "./common"
 
 export function responseValidationFactory(
   possibleResponses: [string, z.ZodTypeAny][],
   defaultResponse?: z.ZodTypeAny,
 ) {
-  // Exploit the natural ordering matching the desired specificity of eg: 404 vs 4xx
-  possibleResponses.sort((x, y) => (x[0] < y[0] ? -1 : 1))
-
-  // TODO: avoid any
-  return async (
-    whenRes: Promise<Res<StatusCode, unknown>>,
-    // biome-ignore lint/suspicious/noExplicitAny: <explanation>
-  ): Promise<any> => {
-    const res = await whenRes
-
-    const json = async () => {
-      const status = res.status
-      const value = await res.json()
-
-      for (const [match, schema] of possibleResponses) {
-        const isMatch =
-          (/^\d+$/.test(match) && String(status) === match) ||
-          (/^\d[xX]{2}$/.test(match) && String(status)[0] === match[0])
-
-        if (isMatch) {
-          return schema.parse(value)
-        }
-      }
-
-      if (defaultResponse) {
-        return defaultResponse.parse(value)
-      }
-
-      // TODO: throw on unmatched response?
-      return value
-    }
-
-    return new Proxy(res, {
-      get(target, prop, receiver) {
-        if (prop === "json") {
-          return json
-        }
-
-        return Reflect.get(target, prop, receiver)
-      },
-    })
-  }
+  return responseValidationFactoryFactory(
+    (schema, value) => {
+      return schema.parse(value)
+    },
+    possibleResponses,
+    defaultResponse,
+  )
 }
